- personalised user id requests implemented

+++ internal/storage/postgres/dish_repository/repository.go
@@ -22,31 +22,31 @@ func NewDishRepository(db postgres.PgxPoolInterface) *DishRepository {
 	}}
 }
 
-func (r *DishRepository) GetDish(id int32) (*dto.DishesResponse, error) {
+func (r *DishRepository) GetDish(dishId, userId int32) (*dto.DishesResponse, error) {
 	var err error
 	dish := &entity.Dish{}
 	goods := &[]dto.GoodsWithAmount{}
 	wg := &sync.WaitGroup{}
 	wg.Add(2)
-	go func(wg *sync.WaitGroup, id int32) {
-		dish, err = r.GetBaseDishInfo(id)
+	go func(wg *sync.WaitGroup, dishId int32, userId int32) {
+		dish, err = r.GetBaseDishInfo(dishId, userId)
 		wg.Done()
-	}(wg, id)
+	}(wg, dishId, userId)
 
-	go func(wg *sync.WaitGroup, id int32) {
-		goods, err = r.GetBaseGoodsInfo(id)
+	go func(wg *sync.WaitGroup, dishId int32, userId int32) {
+		goods, err = r.GetBaseGoodsInfo(dishId, userId)
 		wg.Done()
-	}(wg, id)
+	}(wg, dishId, userId)
 
 	wg.Wait()
 	return &dto.DishesResponse{Dish: *dish, Goods: *goods}, err
 }
 
-func (r *DishRepository) GetAllDishes() (*[]dto.DishesResponse, error) {
+func (r *DishRepository) GetAllDishes(userId int32) (*[]dto.DishesResponse, error) {
 	var dishes []entity.Dish
 	var response []dto.DishesResponse
-	query := fmt.Sprintf("SELECT id, title, description FROM %s", postgres.DishTable)
-	rows, err := r.DB.Query(r.Ctx, query)
+	query := fmt.Sprintf("SELECT id, title, description FROM %s WHERE user_id=$1", postgres.DishTable)
+	rows, err := r.DB.Query(r.Ctx, query, userId)
 	defer rows.Close()
 	for rows.Next() {
 		var dish entity.PgxDish
@@ -59,7 +59,7 @@ func (r *DishRepository) GetAllDishes() (*[]dto.DishesResponse, error) {
 		logger.Error(err)
 	}
 	for _, dish := range dishes {
-		goods, err := r.GetBaseGoodsInfo(dish.Id)
+		goods, err := r.GetBaseGoodsInfo(dish.Id, userId)
 		if err != nil {
 			logger.Error(err)
 			return &response, err
@@ -91,11 +91,11 @@ func (r *DishRepository) CreateDish(dish *entity.Dish) (*entity.Dish, error) {
 	return pgxDish.ToClean(), err
 }
 
-func (r *DishRepository) UpdateDish(dish *entity.Dish, id int32) (*dto.DishesResponse, error) {
+func (r *DishRepository) UpdateDish(dish *entity.Dish, dishId, userId int32) (*dto.DishesResponse, error) {
 	var pgxDish entity.PgxDish
-	query := fmt.Sprintf("UPDATE %s SET updated_at=now(), title=$1, description=$2 WHERE id=$3 RETURNING id, title, description", postgres.DishTable)
+	query := fmt.Sprintf("UPDATE %s SET updated_at=now(), title=$1, description=$2 WHERE id=$3 AND user_id=$4 RETURNING id, title, description", postgres.DishTable)
 	err := r.DB.BeginTxFunc(r.Ctx, pgx.TxOptions{}, func(tx pgx.Tx) error {
-		rows, err := r.DB.Query(r.Ctx, query, dish.Title, dish.Description, id)
+		rows, err := r.DB.Query(r.Ctx, query, dish.Title, dish.Description, dishId, userId)
 		defer rows.Close()
 		for rows.Next() {
 			if err = rows.Scan(&pgxDish.Id, &pgxDish.Title, &pgxDish.Description); err != nil {
@@ -107,17 +107,17 @@ func (r *DishRepository) UpdateDish(dish *entity.Dish, id int32) (*dto.DishesRes
 	if err != nil {
 		logger.Error(err)
 	}
-	goods, err := r.GetBaseGoodsInfo(id)
+	goods, err := r.GetBaseGoodsInfo(dishId, userId)
 	return &dto.DishesResponse{
 		Dish:  *pgxDish.ToClean(),
 		Goods: *goods,
 	}, err
 }
 
-func (r *DishRepository) DeleteDish(id int32) error {
-	query := fmt.Sprintf("DELETE FROM %s WHERE id=$1", postgres.DishTable)
+func (r *DishRepository) DeleteDish(dishId, userId int32) error {
+	query := fmt.Sprintf("DELETE FROM %s WHERE id=$1 AND user_id=$2", postgres.DishTable)
 	err := r.DB.BeginTxFunc(r.Ctx, pgx.TxOptions{}, func(tx pgx.Tx) error {
-		_, err := r.DB.Exec(r.Ctx, query, id)
+		_, err := r.DB.Exec(r.Ctx, query, dishId, userId)
 		return err
 	})
 	if err != nil {
