+++ internal/storage/postgres/goods_repository/repository.go
@@ -20,10 +20,10 @@ func NewGoodsRepository(db postgres.PgxPoolInterface) *GoodsRepository {
 	}}
 }
 
-func (r *GoodsRepository) GetGoods(id int32) (*entity.Goods, error) {
+func (r *GoodsRepository) GetGoods(goodsId, userId int32) (*entity.Goods, error) {
 	var good entity.PgxGoods
-	query := fmt.Sprintf("SELECT id, title, description FROM %s WHERE id=$1", postgres.GoodsTable)
-	rows, err := r.DB.Query(r.Ctx, query, id)
+	query := fmt.Sprintf("SELECT id, title, description FROM %s WHERE id=$1 AND user_id=$2", postgres.GoodsTable)
+	rows, err := r.DB.Query(r.Ctx, query, goodsId, userId)
 	defer rows.Close()
 	for rows.Next() {
 		if err = rows.Scan(&good.Id, &good.Title, &good.Description); err != nil {
@@ -36,10 +36,10 @@ func (r *GoodsRepository) GetGoods(id int32) (*entity.Goods, error) {
 	return good.ToClean(), err
 }
 
-func (r *GoodsRepository) GetAllGoods() (*[]entity.Goods, error) {
+func (r *GoodsRepository) GetAllGoods(userId int32) (*[]entity.Goods, error) {
 	var goods []entity.Goods
-	query := fmt.Sprintf("SELECT id, title, description FROM %s", postgres.GoodsTable)
-	rows, err := r.DB.Query(r.Ctx, query)
+	query := fmt.Sprintf("SELECT id, title, description FROM %s WHERE user_id=$1", postgres.GoodsTable)
+	rows, err := r.DB.Query(r.Ctx, query, userId)
 	defer rows.Close()
 	for rows.Next() {
 		var good entity.PgxGoods
@@ -73,11 +73,12 @@ func (r *GoodsRepository) CreateGoods(good *entity.Goods) (*entity.Goods, error)
 	return pgxGoods.ToClean(), err
 }
 
-func (r *GoodsRepository) UpdateGoods(good *entity.Goods, id int32) (*entity.Goods, error) {
+func (r *GoodsRepository) UpdateGoods(good *entity.Goods, goodsId, userId int32) (*entity.Goods, error) {
 	var pgxGoods entity.PgxGoods
-	query := fmt.Sprintf("UPDATE %s SET updated_at=now(), title=$1, description=$2 WHERE id=$3 RETURNING id, title, description", postgres.GoodsTable)
+	query := fmt.Sprintf("UPDATE %s SET updated_at=now(), title=$1, description=$2 "+
+		"WHERE id=$3 AND user_id=$4 RETURNING id, title, description", postgres.GoodsTable)
 	err := r.DB.BeginTxFunc(r.Ctx, pgx.TxOptions{}, func(tx pgx.Tx) error {
-		rows, err := r.DB.Query(r.Ctx, query, good.Title, good.Description, id)
+		rows, err := r.DB.Query(r.Ctx, query, good.Title, good.Description, goodsId, userId)
 		defer rows.Close()
 		for rows.Next() {
 			if err = rows.Scan(&pgxGoods.Id, &pgxGoods.Title, &pgxGoods.Description); err != nil {
@@ -92,10 +93,10 @@ func (r *GoodsRepository) UpdateGoods(good *entity.Goods, id int32) (*entity.Goo
 	return pgxGoods.ToClean(), err
 }
 
-func (r *GoodsRepository) DeleteGoods(id int32) error {
-	query := fmt.Sprintf("DELETE FROM %s WHERE id=$1", postgres.GoodsTable)
+func (r *GoodsRepository) DeleteGoods(goodsId, userId int32) error {
+	query := fmt.Sprintf("DELETE FROM %s WHERE id=$1 and user_id=$2", postgres.GoodsTable)
 	err := r.DB.BeginTxFunc(r.Ctx, pgx.TxOptions{}, func(tx pgx.Tx) error {
-		_, err := r.DB.Exec(r.Ctx, query, id)
+		_, err := r.DB.Exec(r.Ctx, query, goodsId, userId)
 		return err
 	})
 	if err != nil {
